refactor: rename client.svelte.ts to client.ts across all workspace folders
None of these files use Svelte runes—they're plain TypeScript modules.
The .svelte.ts extension was misleading and inconsistent with zhongwen
which already used client.ts.
