+++ apps/fuji/src/lib/workspace/client.ts
