+++ apps/fuji/src/lib/workspace/index.ts
@@ -1,3 +1,3 @@
-export { default } from './client.svelte';
+export { default } from './client';
 export { type Entry, EntryId, fujiWorkspace } from './definition';
 export { createFuji } from './workspace';
