+++ apps/opensidian/src/lib/workspace/client.ts
