+++ apps/opensidian/src/lib/workspace/index.ts
@@ -1,3 +1,3 @@
-export { workspace, auth, fs, bash } from './client.svelte';
+export { workspace, auth, fs, bash } from './client';
 export { opensidianDefinition } from './definition';
 export { createOpensidian } from './workspace';
