+++ apps/skills/src/lib/workspace/client.ts
