+++ apps/skills/src/lib/workspace/index.ts
@@ -1,3 +1,3 @@
-export { ws, fs, bash } from './client.svelte';
+export { ws, fs, bash } from './client';
 export { skillsEditorDefinition } from './definition';
 export { createSkillsEditor } from './workspace';
