+++ apps/tab-manager/src/lib/workspace/client.ts
