+++ apps/whispering/src/lib/workspace/client.ts
