+++ apps/whispering/src/lib/workspace/index.ts
@@ -1,3 +1,3 @@
-export { default } from './client.svelte';
+export { default } from './client';
 export { whisperingDefinition } from './definition';
 export { createWhispering } from './workspace';
